client/requirement.txt
opencv-python
face_recognition
mysql.connector
requests
